requirements.txt
# Core Flask AppBuilder and dependencies
Flask-AppBuilder==4.3.11
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.0.1
Flask-Admin==2.0.0a4
WTForms==3.0.1

# Database
SQLAlchemy==1.4.41
Alembic==1.11.3

# Authentication & Security
Flask-Security-Too==5.1.2
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1

# API & HTTP
requests==2.31.0
openai==1.3.0
httpx==0.24.1

# Data Processing & Validation
marshmallow==3.19.0
marshmallow-sqlalchemy==0.26.1
python-dotenv==1.0.0

# File Handling
openpyxl==3.1.2

# Development & Testing
pytest==7.3.1
pytest-flask==1.2.0
pytest-cov==4.0.0
black==23.3.0
flake8==6.0.0

# Production
gunicorn==20.1.0
python-decouple==3.8